// src/types/Pokemon.ts
export interface Pokemon {
    abilities: Ability[]
    base_experience: number
    forms: ResourceItem[]
    game_indices: Index[]
    height: number
    held_items: HeldItem[]
    id: number
    is_default: boolean
    location_area_encounters: string
    moves: Mfe[]
    name: string
    order: number
    past_abilities: any[]
    past_types: any[]
    species: ResourceItem
    sprites: Sprites
    stats: Stat[]
    types: Type[]
    weight: number
}

export interface Ability {
    ability: ResourceItem
    is_hidden: boolean
    slot: number
}

export interface Index {
    game_index: number
    version: ResourceItem
}

export interface HeldItem {
    item: ResourceItem
    version_details: VersionDetail[]
}

export interface VersionDetail {
    rarity: number
    version: ResourceItem
}

export interface Mfe {
    move: ResourceItem
    version_group_details: VersionGroupDetail[]
}

export interface VersionGroupDetail {
    level_learned_at: number
    move_learn_method: ResourceItem
    version_group: ResourceItem
}

export interface Sprites {
    back_default: string
    back_female: any
    back_shiny: string
    back_shiny_female: any
    front_default: string
    front_female: any
    front_shiny: string
    front_shiny_female: any
    other: Other
    versions: Versions
}

export interface Other {
    dream_world: DreamWorld
    home: Home
    "official-artwork": OfficialArtwork
    showdown: Showdown
}
// ...
export interface Versions {
    "generation-i": GenerationI
    "generation-ii": GenerationIi
    "generation-iii": GenerationIii
    "generation-iv": GenerationIv
    "generation-v": GenerationV
    "generation-vi": GenerationVi
    "generation-vii": GenerationVii
    "generation-viii": GenerationViii
}
// ...
export interface Stat {
    base_stat: number
    effort: number
    stat: ResourceItem
}

export interface Type {
    slot: number
    type: ResourceItem
}